core-services/egov-rn-service/src/main/java/org/egov/rn/repository/ServiceRequestRepository.java
package org.egov.rn.repository;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.SerializationFeature;
import org.springframework.stereotype.Service;
import org.springframework.web.client.RestTemplate;

import java.util.Map;

@Service
public class ServiceRequestRepository {
    private ObjectMapper objectMapper;
    private RestTemplate restTemplate;

    public ServiceRequestRepository(ObjectMapper objectMapper, RestTemplate restTemplate) {
        this.objectMapper = objectMapper;
        this.restTemplate = restTemplate;
    }


    public Object fetchResult(StringBuilder uri, Object request) {
        objectMapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
        return restTemplate.postForObject(uri.toString(), request, Map.class);
    }
}
